finished Course API

+++ EasyGradeManager/Controllers/API/CoursesController.cs
@@ -1,12 +1,10 @@
 ﻿using EasyGradeManager.Models;
 using System.Collections.Generic;
-using System.Data.Entity;
-using System.Data.Entity.Infrastructure;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using static EasyGradeManager.Static.Authorize;
+using static System.Data.Entity.EntityState;
 
 namespace EasyGradeManager.Controllers.API
 {
@@ -36,67 +34,55 @@ namespace EasyGradeManager.Controllers.API
             return Ok(GetAccessRole(authorizedUser, course) != null ? new CourseDetailDTO(course) : new CourseListDTO(course));
         }
 
-        //TODO implement
-        public IHttpActionResult PutCourse(int id, Course course)
+        public IHttpActionResult PutCourse(int id, CourseDetailDTO courseDTO)
         {
-            if (!ModelState.IsValid)
-            {
+            User authorizedUser = GetAuthorizedUser(Request.Headers.GetCookies("user").FirstOrDefault());
+            if (authorizedUser == null)
+                return Unauthorized();
+            Course course = db.Courses.Find(id);
+            if (courseDTO == null || course == null || !ModelState.IsValid || id != courseDTO.Id)
                 return BadRequest(ModelState);
-            }
-
-            if (id != course.Id)
-            {
+            if (!"Teacher".Equals(GetAccessRole(authorizedUser, course)))
+                return Unauthorized();
+            if (!courseDTO.Validate(course))
                 return BadRequest();
-            }
-
-            db.Entry(course).State = EntityState.Modified;
-
-            try
-            {
-                db.SaveChanges();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!CourseExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return StatusCode(HttpStatusCode.NoContent);
+            courseDTO.Update(course);
+            string error = db.Update(course, Modified);
+            if (error != null)
+                return BadRequest(error);
+            return Redirect("https://" + Request.RequestUri.Host + ":" + Request.RequestUri.Port + "/Courses/" + authorizedUser.Id);
         }
 
-        //TODO implement
-        public IHttpActionResult PostCourse(Course course)
+        public IHttpActionResult PostCourse(CourseDetailDTO courseDTO)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
+            User authorizedUser = GetAuthorizedUser(Request.Headers.GetCookies("user").FirstOrDefault());
+            if (authorizedUser == null || authorizedUser.GetTeacher() == null)
+                return Unauthorized();
+            if (!ModelState.IsValid || !courseDTO.Validate(null))
+                return BadRequest();
+            Course course = courseDTO.Create(authorizedUser.GetTeacher().Id);
             db.Courses.Add(course);
-            db.SaveChanges();
-
-            return CreatedAtRoute("DefaultApi", new { id = course.Id }, course);
+            string error = db.Update(course, Added);
+            if (error != null)
+                return BadRequest(error);
+            return Redirect("https://" + Request.RequestUri.Host + ":" + Request.RequestUri.Port + "/Courses/" + course.Id);
         }
 
-        //TODO implement
         public IHttpActionResult DeleteCourse(int id)
         {
+            User authorizedUser = GetAuthorizedUser(Request.Headers.GetCookies("user").FirstOrDefault());
+            if (authorizedUser == null)
+                return Unauthorized();
             Course course = db.Courses.Find(id);
             if (course == null)
-            {
                 return NotFound();
-            }
-
+            if (!"Teacher".Equals(GetAccessRole(authorizedUser, course)))
+                return Unauthorized();
             db.Courses.Remove(course);
-            db.SaveChanges();
-
-            return Ok(course);
+            string error = db.Update(course, Deleted);
+            if (error != null)
+                return BadRequest(error);
+            return Redirect("https://" + Request.RequestUri.Host + ":" + Request.RequestUri.Port + "/Users/" + authorizedUser.Id);
         }
 
         protected override void Dispose(bool disposing)
@@ -106,11 +92,6 @@ namespace EasyGradeManager.Controllers.API
                 db.Dispose();
             }
             base.Dispose(disposing);
-        }
-
-        private bool CourseExists(int id)
-        {
-            return db.Courses.Count(e => e.Id == id) > 0;
         }
     }
 }
